start HitRequest Container for more readability

+++ src/ChickenAPI.Game/Data/AccessLayer/Battle/IDamageAlgorithm.cs
@@ -5,11 +5,11 @@ namespace ChickenAPI.Game.Data.AccessLayer.Battle
     public interface IDamageAlgorithm
     {
         /// <summary>
-        /// returns the damages when one entity strikes another based on their stats
+        /// Returns the damages when one entity strikes another based on their stats
         /// </summary>
         /// <param name="entity"></param>
         /// <param name="targetEntity"></param>
         /// <returns></returns>
-        ushort GenerateDamage(IBattleEntity entity, IBattleEntity targetEntity);
+        uint GenerateDamage(IBattleEntity entity, IBattleEntity targetEntity);
     }
 }
